@@ -22,8 +22,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     ndkVersion '25.2.9519653'
 }
@@ -37,7 +37,13 @@ dependencies {
     implementation project(path: ':H2CO3Core')
     implementation project(path: ':H2CO3Resources')
     implementation project(path: ':H2CO3Boat')
+
+    // Google
     implementation 'com.google.code.gson:gson:2.10.1'
+    implementation 'com.google.android.material:material:1.9.0'
+    implementation 'com.google.dagger:hilt-android:2.48'
+
+    // AndroidX
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.navigation:navigation-fragment:2.6.0'
     implementation 'androidx.navigation:navigation-ui:2.6.0'
@@ -49,19 +55,23 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.6.1'
     implementation 'androidx.preference:preference-ktx:1.2.0'
     implementation 'androidx.legacy:legacy-support-v13:1.0.0'
-    implementation 'com.afollestad.material-dialogs:commons:0.9.6.0'
-    implementation 'com.android.billingclient:billing:5.1.0'
-    implementation 'com.github.bumptech.glide:glide:4.15.1'
-    implementation 'com.google.android.material:material:1.9.0'
-    implementation 'com.google.dagger:hilt-android:2.47'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.test.ext:junit:1.1.5'
+    implementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'androidx.preference:preference:1.2.0'
+    implementation 'androidx.webkit:webkit:1.7.0-alpha03'
 
+    // Apache Commons
     implementation 'org.apache.commons:commons-text:1.10.0'
     implementation 'commons-io:commons-io:20030203.000550'
 
+    // Activity Permissions
     implementation 'com.github.getActivity:XXPermissions:16.8'
 
+    // Custom Activity On Crash
     implementation 'cat.ereza:customactivityoncrash:2.4.0'
 
+    // Rikka
     implementation 'dev.rikka.rikkax.appcompat:appcompat:1.6.1'
     implementation 'dev.rikka.rikkax.compatibility:compatibility:2.0.0'
     implementation 'dev.rikka.rikkax.material:material:2.7.0'
@@ -77,19 +87,28 @@ dependencies {
     implementation 'dev.rikka.rikkax.lifecycle:lifecycle-shared-viewmodel:1.0.1'
     implementation 'dev.rikka.rikkax.lifecycle:lifecycle-viewmodel-lazy:2.0.0'
 
+    // Volley
     implementation 'com.android.volley:volley:1.2.1'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'androidx.preference:preference:1.2.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
 
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    testImplementation 'junit:junit:4.13.2'
+    // Fastjson
+    implementation 'com.alibaba:fastjson:2.0.40'
 
-    implementation 'com.alibaba:fastjson:2.0.36'
-    //implementation 'com.google.code.gson:gson:2.8.2'
+    // OkHttp
     implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.11'
+
+    // HoloColorPicker
     implementation 'com.larswerkman:HoloColorPicker:1.5'
-    implementation 'androidx.webkit:webkit:1.7.0-alpha03'
+
+    // GreenRobot EventBus
     implementation "org.greenrobot:eventbus:3.3.1"
+
+    // Commons IO
+    implementation 'commons-io:commons-io:2.11.0'
+
+    // Billing Client
+    implementation 'com.android.billingclient:billing:5.1.0'
+
+    // Glide
+    implementation 'com.github.bumptech.glide:glide:4.15.1'
+    implementation project(path: ':H2CO3Dialog')
 }
