@@ -20,8 +20,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 }
 
@@ -47,7 +47,7 @@ dependencies {
     implementation 'com.android.billingclient:billing:5.1.0'
     implementation 'com.github.bumptech.glide:glide:4.15.1'
     implementation 'com.google.android.material:material:1.9.0'
-    implementation 'com.google.dagger:hilt-android:2.47'
+    implementation 'com.google.dagger:hilt-android:2.48'
 
     implementation 'org.apache.commons:commons-text:1.10.0'
 
@@ -79,7 +79,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     testImplementation 'junit:junit:4.13.2'
 
-    implementation 'com.alibaba:fastjson:2.0.36'
+    implementation 'com.alibaba:fastjson:2.0.40'
     //implementation 'com.google.code.gson:gson:2.8.2'
     implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.11'
     implementation 'com.larswerkman:HoloColorPicker:1.5'
