@@ -18,10 +18,11 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 }
+
 configurations.configureEach {
     exclude group: 'androidx.appcompat', module: 'appcompat'
 }
@@ -29,7 +30,9 @@ configurations.configureEach {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(path: ':H2CO3Core')
-    implementation 'com.google.code.gson:gson:2.10.1'
+
+    // AndroidX
+    implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.navigation:navigation-fragment:2.6.0'
     implementation 'androidx.navigation:navigation-ui:2.6.0'
@@ -41,19 +44,16 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.6.1'
     implementation 'androidx.preference:preference-ktx:1.2.0'
     implementation 'androidx.legacy:legacy-support-v13:1.0.0'
-    implementation 'com.afollestad.material-dialogs:commons:0.9.6.0'
-    implementation 'com.android.billingclient:billing:5.1.0'
+
+    // Third-party libraries
+    implementation 'com.google.code.gson:gson:2.10.1'
     implementation 'com.github.bumptech.glide:glide:4.15.1'
     implementation 'com.google.android.material:material:1.9.0'
-    implementation 'com.google.dagger:hilt-android:2.47'
-
+    implementation 'com.google.dagger:hilt-android:2.48'
     implementation 'org.apache.commons:commons-text:1.10.0'
     implementation 'commons-io:commons-io:20030203.000550'
-
     implementation 'com.github.getActivity:XXPermissions:16.8'
-
     implementation 'cat.ereza:customactivityoncrash:2.4.0'
-
     implementation 'dev.rikka.rikkax.appcompat:appcompat:1.6.1'
     implementation 'dev.rikka.rikkax.compatibility:compatibility:2.0.0'
     implementation 'dev.rikka.rikkax.material:material:2.7.0'
@@ -68,23 +68,18 @@ dependencies {
     implementation 'dev.rikka.rikkax.lifecycle:lifecycle-resource-livedata:1.0.1'
     implementation 'dev.rikka.rikkax.lifecycle:lifecycle-shared-viewmodel:1.0.1'
     implementation 'dev.rikka.rikkax.lifecycle:lifecycle-viewmodel-lazy:2.0.0'
-
     implementation 'com.android.volley:volley:1.2.1'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'androidx.preference:preference:1.2.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.webkit:webkit:1.7.0-alpha03'
+    implementation "org.greenrobot:eventbus:3.3.1"
+    implementation 'commons-codec:commons-codec:1.16.0'
+    implementation project(path: ':ultimatebarx')
 
+    // Testing dependencies
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     testImplementation 'junit:junit:4.13.2'
-
-    implementation 'com.alibaba:fastjson:2.0.36'
-    //implementation 'com.google.code.gson:gson:2.8.2'
+    implementation 'com.alibaba:fastjson:2.0.40'
     implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.11'
     implementation 'com.larswerkman:HoloColorPicker:1.5'
-    implementation 'androidx.webkit:webkit:1.7.0-alpha03'
-    implementation "org.greenrobot:eventbus:3.3.1"
-    // https://mvnrepository.com/artifact/commons-codec/commons-codec
-    implementation 'commons-codec:commons-codec:1.16.0'
-
 }
