@@ -1,27 +1,26 @@
-package org.koishi.launcher.h2co3.ui.home;
+package org.koishi.launcher.h2co3.ui.fragment.directory;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+
 import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 
 import org.koishi.launcher.h2co3.R;
 import org.koishi.launcher.h2co3.resources.component.H2CO3Fragment;
 
-public class HomeFragment extends H2CO3Fragment {
+public class DirectoryFragment extends H2CO3Fragment {
+
 
     public View onCreateView(@NonNull LayoutInflater inflater,
-            ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_home, container, false);
+                             ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_directory, container, false);
 
         return view;
     }
 
-@Override
+    @Override
     public void onDestroyView() {
         super.onDestroyView();
     }
