@@ -1,4 +1,4 @@
-package org.koishi.launcher.h2co3.ui.manage;
+package org.koishi.launcher.h2co3.ui.fragment.home;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,16 +10,16 @@ import androidx.annotation.NonNull;
 import org.koishi.launcher.h2co3.R;
 import org.koishi.launcher.h2co3.resources.component.H2CO3Fragment;
 
-public class ManageFragment extends H2CO3Fragment {
+public class HomeFragment extends H2CO3Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_manage, container, false);
+            ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_home, container, false);
 
         return view;
     }
 
-    @Override
+@Override
     public void onDestroyView() {
         super.onDestroyView();
     }
