@@ -1,4 +1,4 @@
-package org.koishi.launcher.h2co3.ui.directory;
+package org.koishi.launcher.h2co3.ui.fragment.manage;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,12 +10,11 @@ import androidx.annotation.NonNull;
 import org.koishi.launcher.h2co3.R;
 import org.koishi.launcher.h2co3.resources.component.H2CO3Fragment;
 
-public class DirectoryFragment extends H2CO3Fragment {
-
+public class ManageFragment extends H2CO3Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_directory, container, false);
+        View view = inflater.inflate(R.layout.fragment_manage, container, false);
 
         return view;
     }
