@@ -7,7 +7,7 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import org.koishi.launcher.h2co3.resources.R;
+import org.koishi.launcher.h2co3.dialog.R;
 import org.koishi.launcher.h2co3.dialog.list.events.ListItemClick;
 
 import java.util.ArrayList;
